aag_cloudwatcher_ng: make getRefreshPeriod() use Wheater interface property
The WEATHER_UPDATE property is defined / deleted when connection status changes, so we just retrieve it from UpdatePeriodN[0].value

+++ indi-aagcloudwatcher-ng/indi_aagcloudwatcher_ng.h
@@ -51,7 +51,7 @@ class AAGCloudWatcher : public INDI::Weather
 
     virtual const char *getDefaultName() override;
     bool sendData();
-    int getRefreshPeriod();
+    float getRefreshPeriod();
     float getLastReadPeriod();
     bool heatingAlgorithm();
